+++ eslint.config.mjs
@@ -48,6 +48,7 @@ const ignores = [
 
 // These rules apply to all files
 const sharedRules = {
+  'internal/no-object-rest-spread-in-worklet': 'error',
   'import/default': 'off',
   'import/extensions': 'off',
   'import/named': 'off',
@@ -180,6 +181,7 @@ const testRules = {
 
 // These plugins apply to all files
 const sharedPlugins = {
+  internal: internalPlugin,
   'simple-import-sort': eslintSimpleImportSort,
 };