+++ packages/web-visualization/src/chart/line/ReferenceLine.tsx
@@ -169,6 +169,10 @@ export type ReferenceLineProps = (HorizontalReferenceLineProps | VerticalReferen
      */
     root?: string;
     /**
+     * Custom class name for the line path.
+     */
+    line?: string;
+    /**
      * Custom class name for the text label.
      */
     label?: string;
@@ -181,6 +185,10 @@ export type ReferenceLineProps = (HorizontalReferenceLineProps | VerticalReferen
      * Custom styles for the root element.
      */
     root?: React.CSSProperties;
+    /**
+     * Custom styles for the line path.
+     */
+    line?: React.CSSProperties;
     /**
      * Custom styles for the text label.
      */
@@ -245,9 +253,11 @@ export const ReferenceLine = memo<ReferenceLineProps>(
         <g className={rootClassName} data-testid={testID} style={rootStyle}>
           <LineComponent
             animate={false}
+            className={classNames?.line}
             d={`M${drawingArea.x},${yPixel} L${drawingArea.x + drawingArea.width},${yPixel}`}
             stroke={stroke}
             strokeOpacity={opacity}
+            style={styles?.line}
           />
           {label && (
             <LabelComponent
@@ -298,9 +308,11 @@ export const ReferenceLine = memo<ReferenceLineProps>(
         <g className={rootClassName} data-testid={testID} style={rootStyle}>
           <LineComponent
             animate={false}
+            className={classNames?.line}
             d={`M${xPixel},${drawingArea.y} L${xPixel},${drawingArea.y + drawingArea.height}`}
             stroke={stroke}
             strokeOpacity={opacity}
+            style={styles?.line}
           />
           {label && (
             <LabelComponent