+++ packages/web-visualization/CHANGELOG.md
@@ -8,6 +8,13 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.4.0-beta.21 (3/2/2026 PST)
+
+#### 🚀 Updates
+
+- Fix styles props on Scrubber. [[#463](https://github.com/coinbase/cds/pull/463)]
+- Fix text elevation background. [[#463](https://github.com/coinbase/cds/pull/463)]
+
 ## 3.4.0-beta.20 (2/27/2026 PST)
 
 #### 🚀 Updates